starting to fix auth errors
